Move secret store related test to new subclass.
